Move rest layer tests for filters to their own file
